# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CRM module for Uralintern
  description: API documentation for Uralintern CRM Module
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: kayakto server url
security:
  - BearerAuth: []
...
paths:
  /users/me:
    get:
      tags:
        - user-info-controller
      operationId: getCurrentUser
      description: "Получение данных пользователя"
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    put:
      tags:
        - user-info-controller
      operationId: updateCurrentUser
      description: "Обновление данных пользователя"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /users/all-managers:
    get:
      tags:
        - user-info-controller
      operationId: getAllManagers
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  responses:
      OkResponse:
        description: OK, запрос выполнен успешно
        content:
          '*/*':
            schema:
              type: boolean
      BadRequestResponse:
        description: Bad Request, запрос неверный со стороны клиента
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      NotFoundResponse:
        description: Not Found, такого элемента не существует
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      InternalServerErrorResponse:
        description: Internal Server Error, запрос вызвал ошибку на сервере
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "User with email email123@gmail.com not found"
    UserUpdateRequest:
      required:
        - firstName
        - lastName
        - telegramUrl
        - vkUrl
      type: object
      properties:
        firstName:
          type: string
          example: "Николай"
        lastName:
          type: string
          example: "Путин"
        surname:
          type: string
          example: "Евгеньевич"
        telegramUrl:
          type: string
          example: "https://t.me/kayakto"
        vkUrl:
          type: string
          example: "https://vk.com/kayakto"
        competencies:
          type: string
          example: "Spring Boot, Gradle, git, Java, Docker"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT